Move ReflectionHelpers to Shared
