+++ tests/Agent/IntegrationTests/ApplicationHelperLibraries/ConsoleMultiFunctionApplicationHelpers/NetStandardLibraries/FasterEventHarvest.cs
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 
-using NewRelic.Agent.IntegrationTests.ApplicationHelpers;
+using NewRelic.Agent.IntegrationTests.Shared.ReflectionHelpers;
 using NewRelic.Api.Agent;
 
 namespace MultiFunctionApplicationHelpers.NetStandardLibraries
