+++ tests/Agent/IntegrationTests/ApplicationHelperLibraries/ConsoleMultiFunctionApplicationHelpers/NetStandardLibraries/ProcessRunner.cs
@@ -2,11 +2,10 @@
 // SPDX-License-Identifier: Apache-2.0
 
 
-using NewRelic.Agent.IntegrationTests.ApplicationHelpers;
+using NewRelic.Agent.IntegrationTests.Shared.ReflectionHelpers;
 using System;
 using System.Diagnostics;
 using System.Text;
-using System.Threading;
 
 namespace MultiFunctionApplicationHelpers.NetStandardLibraries
 {
