+++ tests/Agent/IntegrationTests/ApplicationHelperLibraries/ConsoleMultiFunctionApplicationHelpers/NetStandardLibraries/W3C/W3CTestService.cs
@@ -3,7 +3,7 @@
 
 
 using MultiFunctionApplicationHelpers.NetStandardLibraries.Owin;
-using NewRelic.Agent.IntegrationTests.ApplicationHelpers;
+using NewRelic.Agent.IntegrationTests.Shared.ReflectionHelpers;
 using System;
 
 namespace MultiFunctionApplicationHelpers.NetStandardLibraries.W3C
