+++ tests/Agent/IntegrationTests/ApplicationHelperLibraries/NetStandardTestLibrary/NetStandardTestLibrary.csproj
@@ -5,7 +5,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\ReflectionHelpers\ReflectionHelpers.csproj" />
+    <ProjectReference Include="..\..\Shared\Shared.csproj" />
   </ItemGroup>
 
 </Project>
