+++ tests/Agent/IntegrationTests/Applications/ConsoleInstrumentationLoader/Program.cs
@@ -5,7 +5,7 @@
 using NewRelic.Api.Agent;
 using System;
 using System.Runtime.CompilerServices;
-using NewRelic.Agent.IntegrationTests.ApplicationHelpers;
+using NewRelic.Agent.IntegrationTests.Shared.ReflectionHelpers;
 
 namespace ConsoleInstrumentationLoader
 {
