+++ tests/Agent/IntegrationTests/SharedApplications/ConsoleMultiFunctionApplicationFW/ConsoleMultiFunctionApplicationFW.csproj
@@ -29,8 +29,8 @@
     <ProjectReference Include="..\..\ApplicationHelperLibraries\ConsoleMultiFunctionApplicationHelpers\MultiFunctionApplicationHelpers.csproj" />
     <ProjectReference Include="..\..\ApplicationHelperLibraries\IntegrationTestFrameworkHelpers\IntegrationTestFrameworkHelpers.csproj" />
     <ProjectReference Include="..\..\ApplicationHelperLibraries\NetStandardTestLibrary\NetStandardTestLibrary.csproj" />
-    <ProjectReference Include="..\..\ApplicationHelperLibraries\ReflectionHelpers\ReflectionHelpers.csproj" />
     <ProjectReference Include="..\..\IntegrationTestHelpers\IntegrationTestHelpers.csproj" />
+    <ProjectReference Include="..\..\Shared\Shared.csproj" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Properties\" />
